ops: centralize mm length conversion
